+++ src/controller/cableLengthController.cpp
@@ -26,6 +26,7 @@
     description: A Cable length controller for joint position targets using PD control
 */
 
+
 #include <type_traits>
 #include <controller_interface/controller.h>
 #include <hardware_interface/joint_command_interface.h>
@@ -33,50 +34,18 @@
 #include "kindyn/robot.hpp"
 #include "kindyn/controller/cardsflow_state_interface.hpp"
 #include <roboy_simulation_msgs/ControllerType.h>
+#include <math.h>
+#include <cmath>
 #include <std_msgs/Float32.h>
 #include <roboy_control_msgs/SetControllerParameters.h>
 
+# define M_2PI 2*M_PI
+
 using namespace std;
-
-static const double     _PI= 3.1415926535897932384626433832795028841971693993751058209749445923078164062862089986280348;
-static const double _TWO_PI= 6.2831853071795864769252867665590057683943387987502116419498891846156328125724179972560696;
-
-double mod(double x, double y)
-{
-    double m= x - y * floor(x/y);
-    // handle boundary cases resulted from floating-point cut off:
-    if (y > 0)              // modulo range: [0..y)
-    {
-        if (m >= y)           // Mod(-1e-16             , 360.    ): m= 360.
-            return 0;
-
-        if (m < 0)
-        {
-            if (y+m == y)
-                return 0;    // just in case...
-            else
-                return y+m;  // Mod(106.81415022205296 , _TWO_PI ): m= -1.421e-14
-        }
-    }
-    else                    // modulo range: (y..0]
-    {
-        if (m <= y)           // Mod(1e-16              , -360.   ): m= -360.
-            return 0;
-
-        if (m > 0)
-        {
-            if (y+m == y)
-                return 0;    // just in case...
-            else
-                return y+m;  // Mod(-106.81415022205296, -_TWO_PI): m= 1.421e-14
-        }
-    }
-    return m;
-}
 
 double wrap_pos_neg_pi(double angle)
 {
-    return mod(angle + _PI, _TWO_PI) - _PI;
+    return fmod(angle + M_PI, M_2PI) - M_PI;
 }
 
 class CableLengthController : public controller_interface::Controller<hardware_interface::CardsflowCommandInterface> {
@@ -125,7 +94,7 @@ public:
         MatrixXd L = joint.getL();
         double p_error = wrap_pos_neg_pi(q - q_target);
         // we use the joint_index column of the L matrix to calculate the result for this joint only
-        VectorXd ld = L.col(joint_index) * ((*joint.Kd_dl_) * (p_error - p_error_last)/period.toSec() + (*joint.Kp_dl_) * p_error);
+        VectorXd ld = L.col(joint_index) * ((*joint.Kd_) * (p_error - p_error_last)/period.toSec() + (*joint.Kp_) * p_error);
         joint.setMotorCommand(ld);
         p_error_last = p_error;
         last_update = time;
